+++ NEWS.md
@@ -1,5 +1,7 @@
 # testthat (development version)
 
+* `test_path()` now works when called within helper files (#1562).
+
 * `it()` now calls `local_test_context()` so that it behaves more
   similarly to `test_that()` (#1731), and is now exported so that you
   can more easily run BDD tests interactively (#1587)